Corrected several errors in tutorial
There were several errors in the sample run script in how the arguments were being provided to the prompts of the m3fake tool.  These have now been corrected.  The key bug was that that the original script was selecting IINTEGER as the data type instead of REAL which would cause an error when running CMAQ.